refactor code - go mod tidy - file rename to be more explicit
